% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{AP}
\alias{AP}
\title{Associated Press News Articles}
\format{A list elements extracted from a topic model fit to this data
\describe{
  \item{phi}{phi, a matrix with the topic-term distributions}
  \item{theta}{theta, a matrix with the document-topic distributions}
  \item{doc.length}{doc.length, a numeric vector with token counts for each document}
  \item{vocab}{vocab, a character vector containing the terms}
  \item{term.frequency}{term.frequency, a numeric vector of observed term frequencies}
}}
\source{
\url{http://www.cs.princeton.edu/~blei/lda-c/index.html}
}
\usage{
AP
}
\description{
Associated Press News Articles
}
\keyword{datasets}

