% Generated by roxygen2 (4.0.1): do not edit by hand
\name{distance}
\alias{distance}
\title{Collection of distance matrix computation functions.}
\usage{
distance(measure = "JS")
}
\arguments{
  \item{method}{character string. Specifies which distance
  measure to use. 'JS' and 'KL' are currently supported.}
}
\value{
Returns a bivariate function

x <- matrix(rnorm(100, mean = 100), nrow = 5) dist(x)
library(proxy) dist(x, method = distance()) dist(x, method
= distance(measure = "KL"))
}
\description{
The proxy library allows us to pass an arbitrary bivariate
function for distance matrix computations. This function
will return different bivariate functions based on the
input.
}

