% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{Jeopardy}
\alias{Jeopardy}
\title{Jeopardy Questions (including category name and answer)}
\format{A list elements extracted from a topic model fit to this data
\describe{
  \item{phi}{phi, a matrix with the topic-term distributions}
  \item{theta}{theta, a matrix with the document-topic distributions}
  \item{doc.length}{doc.length, a numeric vector with token counts for each document}
  \item{vocab}{vocab, a character vector containing the terms}
  \item{term.frequency}{term.frequency, a numeric vector of observed term frequencies}
}}
\source{
\url{http://www.reddit.com/r/datasets/comments/1uyd0t/200000_jeopardy_questions_in_a_json_file}
}
\usage{
Jeopardy
}
\description{
Jeopardy Questions (including category name and answer)
}
\keyword{datasets}

