% Generated by roxygen2 (4.0.2): do not edit by hand
\name{jsPCA}
\alias{jsPCA}
\title{Dimension reduction via Jensen-Shannon Divergence & Principal Components}
\usage{
jsPCA(phi)
}
\arguments{
  \item{phi}{matrix, with each row containing the
  distribution over terms for a topic, with as many rows as
  there are topics in the model, and as many columns as
  there are terms in the vocabulary.}
}
\description{
Dimension reduction via Jensen-Shannon Divergence &
Principal Components
}

