% Generated by roxygen2 (4.0.1): do not edit by hand
\name{renderVis}
\alias{renderVis}
\title{Create an LDAvis output element}
\usage{
renderVis(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
  \item{expr}{}

  \item{env}{}

  \item{quoted}{}
}
\description{
Shiny server output function customized for animint plots
(similar to \code{shiny::plotOutput} and friends).
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}

