% Generated by roxygen2 (4.0.2): do not edit by hand
\name{serVis}
\alias{serVis}
\title{View/share LDAvis}
\usage{
serVis(json, out.dir = tempfile(), open.browser = interactive(),
  as.gist = FALSE, ...)
}
\arguments{
  \item{json}{character string output from the
  \link{createjson} function.}

  \item{out.dir}{directory to store html/js/json files.}

  \item{open.browser}{Should R open a browser? If yes, this
  function will attempt to create a local file server via
  the servr package. This is necessary since the javascript
  needs to access local files and most browsers will not
  allow this.}

  \item{as.gist}{should the vis be uploaded as a gist? If
  so, make sure your 'github.username' and
  'github.password' are set in \link{options}.}

  \item{...}{arguments passed onto
  \code{gistr::gist_create}}
}
\value{
An invisible object.
}
\description{
Then, if appropriate, the function will prompt your browser
to serve the visualization.
}
\details{
This function will place the necessary html/js/css files
(located in \code{system.file("htmljs", package =
"LDAvis")}) in a directory specified by the argument
\code{out.dir}.
}
\examples{
\dontrun{
help(createJSON, package = "LDAvis")
}
}
\author{
Carson Sievert
}
\seealso{
\link{createJSON}
}

