\name{sortPhi}
\alias{sortPhi}
\title{Sort phi before uploading to runVis()}
\usage{
  sortPhi(phi, ...)
}
\arguments{
  \item{phi}{numeric matrix where column names are tokens
  and rownames are topic labels}
}
\value{
  A matrix with number of columns equal to the vocabulary
  and number of rows equal to the number of topics
}
\description{
  Sort phi before uploading to runVis()
}

