% Generated by roxygen2 (4.0.1): do not edit by hand
\name{visOutput}
\alias{visOutput}
\title{Shiny ui output function}
\usage{
visOutput(outputId)
}
\arguments{
  \item{outputId}{output variable to read the plot from}
}
\description{
Shiny ui output function
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}

