% Generated by roxygen2 (4.0.2): do not edit by hand
\name{newJSON}
\alias{newJSON}
\title{Create the JSON object to read into the javascript visualization}
\usage{
newJSON(phi = matrix(), theta = matrix(), alpha = numeric(),
  beta = numeric(), doc.length = integer(), vocab = character(),
  term.frequency = integer(), R = 30, quiet = FALSE)
}
\arguments{
  \item{phi}{matrix, with each row containing the
  distribution over terms for a topic, with as many rows as
  there are topics in the model, and as many columns as
  there are terms in the vocabulary.}

  \item{theta}{matrix, with each row containing the
  probability distribution over topics for a document, with
  as many rows as there are documents in the corpus, and as
  many columns as there are topics in the model.}

  \item{alpha}{numeric vector with as many elements as
  there are topics in the model, containing the parameters
  of the Dirichlet prior distribution over topics for each
  document.}

  \item{beta}{numeric vector with as many elements as there
  are terms in the vocabulary, containing the parameters of
  the Dirichlet prior distribution over terms for each
  topic.}

  \item{doc.length}{integer vector containing the number of
  tokens in each document in the corpus.}

  \item{vocab}{character vector of the terms in the
  vocabulary (in the same order as the columns of
  \code{phi} and the elements of \code{beta}).}

  \item{term.frequency}{integer vector containing the
  frequency of each term in the vocabulary.}

  \item{R}{integer, the number of terms to display in the
  barcharts of the interactive viz. Default is 30.
  Recommended to be roughly between 10 and 50.}

  \item{quiet}{logical; should the function print progress
  to the screen during computation?}
}
\value{
A string containing JSON content which can be written to a
file or feed into \link{serVis} for easy viewing/sharing.
}
\description{
This function creates the JSON object that feeds the
visualization template. For a more detailed overview, see
\code{vi}
}
\details{
The function first computes the topic frequencies (across
the whole corpus), and then it reorders the topics in
decreasing order of frequency. The main computation is to
loop through the topics and through 101 values of lambda
(0, 0.01, 0.02, .., 1) to compute the R most
\emph{relevant} terms for each topic and value of lambda.
If \code{print.progress = TRUE} progress in this loop
(which can take a minute or two) will print to the screen.
}
\examples{
# This example uses news article data from D = 2246 Associated Press
# articles tokenized and shared by David Blei:
# http://www.cs.princeton.edu/~blei/lda-c/index.html
\dontrun{
# load the AP data:
data(AP, package="LDAvis")

# create the json object:
json <- newJSON(phi = AP$phi, theta = AP$theta, alpha = AP$alpha,
               beta = AP$beta, doc.length = AP$doc.length,
               vocab = AP$vocab, term.frequency = AP$term.frequency,
               R = 30, print.progress = TRUE)
# By default, serVis will attempt to run a local file server
# in a temporary directory (and prompt your browser to open)
serVis(json) # press ESC or Ctrl-C to kill

# You may want to just write the JSON and other dependency files
# to a folder named AP under the working directory
serVis(json, out.dir = 'AP', open.browser = FALSE)
# then you could use a server of your choice
system("cd AP && python -m SimpleHTTPServer", wait = FALSE)
browseURL("http://localhost:8000")

# If you have a GitHub account, you can even publish as a gist
# which allows you to easily share with others!
serVis(json, as.gist = TRUE)
}
}
\references{
Sievert, C. and Shirley, K. (2014) \emph{LDAvis: A Method
for Visualizing and Interpreting Topics}, ACL Workshop on
Interactive Language Learning, Visualization, and
Interfaces.
\url{http://nlp.stanford.edu/events/illvi2014/papers/sievert-illvi2014.pdf}
}
\seealso{
\link{serVis}
}

