% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderVis}
\alias{renderVis}
\title{Create an LDAvis output element}
\usage{
renderVis(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that generates a plot.}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is expr a quoted expression (with \code{quote()})? This is useful if you want to save an expression in a variable.}
}
\description{
Shiny server output function customized for animint plots 
(similar to \code{shiny::plotOutput} and friends).
}
\seealso{
http://shiny.rstudio.com/articles/building-outputs.html
}
