% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShiny.R
\name{runShiny}
\alias{runShiny}
\title{Run shiny/D3 visualization}
\usage{
runShiny(phi, term.frequency, vocab, topic.proportion)
}
\arguments{
\item{phi}{a matrix with W rows, one for each term in the vocabulary, and K 
columns, one for each topic, where each column sums to one. Each column is the 
multinomial distribution over terms for a given topic in an LDA topic model.}

\item{term.frequency}{an integer vector of length W containing the frequency 
of each term in the vocabulary.}

\item{vocab}{a character vector of length W containing the unique terms in 
the corpus.}

\item{topic.proportion}{a numeric vector of length K containing the proportion
of each topic in the corpus.}
}
\description{
This function is deprecated as of version 0.2
}
