% Generated by roxygen2 (4.0.2): do not edit by hand
\name{runVis}
\alias{runVis}
\title{Run shiny/D3 visualization}
\usage{
runVis(...)
}
\arguments{
  \item{...}{Arguments passed to shiny::runApp.}
}
\description{
This function was deprecated and replaced by
\link{runShiny}
}
\seealso{
\link{check.inputs}
}

